requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
ca-certificates=2022.6.15=h033912b_0
certifi=2022.6.15=py38h50d1736_0
colorama=0.4.5=pyhd8ed1ab_0
cycler=0.11.0=pyhd8ed1ab_0
freetype=2.11.0=hd8bbffd_0
giflib=5.2.1=haf1e3a3_0
intel-openmp=2022.0.0=hecd8cb5_3615
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=hca72f7f_0
kiwisolver=1.4.2=py38he9d5cce_0
lcms2=2.12=hf1fd2bf_0
libblas=3.9.0=15_osx64_openblas
libcblas=3.9.0=15_osx64_openblas
libcxx=12.0.0=h2f01273_0
libedit=3.1.20210910=hca72f7f_0
libffi=3.2.1=h0a44026_1007
libgfortran=5.0.0=9_3_0_h6c81a4c_23
libgfortran5=9.3.0=h6c81a4c_23
liblapack=3.9.0=15_osx64_openblas
libopenblas=0.3.20=openmp_hb3cd9ec_0
libpng=1.6.37=ha441bb4_0
libtiff=4.2.0=hdb42f99_1
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
llvm-openmp=14.0.4=ha654fa7_0
lz4-c=1.9.3=h23ab428_1
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py38h1300a51_1
mkl=2022.0.0=hecd8cb5_105
ncurses=6.3=hca72f7f_2
ninja=1.10.2=hecd8cb5_5
ninja-base=1.10.2=haf03e11_5
numpy=1.22.1=py38h9d72dae_0
openssl=1.1.1o=hfe4f2af_0
pillow=9.0.1=py38hde71d04_0
pip=21.2.4=py38hecd8cb5_0
pyparsing=3.0.9=pyhd8ed1ab_0
python=3.8.0=h359304d_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.4.0=py3.8_0
readline=7.0=h1de35cc_5
scikit-learn=1.0.2=py38h37f3bb3_0
scipy=1.8.0=py38hd329d04_1
setuptools=61.2.0=py38hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=hffcf06c_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h5d9f67b_0
torchaudio=0.4.0=py38
torchvision=0.5.0=py38_cpu
tornado=6.1=py38hed1de0f_3
tqdm=4.64.0=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=hca72f7f_1
zlib=1.2.12=h4dc903c_2
zstd=1.5.2=hcb37349_0
